refactor: extract sidebar navigation to separate component
- Created SidebarNavigation component with all menu logic
- Moved menuSections configuration to SidebarNavigation
- Added utility functions getDataScopeInfo and getCurrentPageTitle
- Updated App.tsx to use the new SidebarNavigation component
- Removed duplicated sidebar JSX and menu logic from App.tsx
- Maintained all existing functionality while improving code organization
